Small fixes, refactoring, cleanup
