+++ Assets/Scripts/Classes/Citybuilding/Buildings/MainCamp/Upgrades/Level2.cs
@@ -14,7 +14,7 @@ namespace Classes.Citybuilding.Buildings.MainCamp.Upgrades
             };
             ModelName = "MainCamp2";
         }
-        
+
         public override void ApplySideEffects(Simulation simulation, Building building)
         {
             base.ApplySideEffects(simulation, building);
