+++ Assets/Scripts/Classes/Citybuilding/Buildings/MainCamp/Upgrades/Level3.cs
@@ -14,7 +14,7 @@ namespace Classes.Citybuilding.Buildings.MainCamp.Upgrades
             };
             ModelName = "MainCamp3";
         }
-        
+
         public override void ApplySideEffects(Simulation simulation, Building building)
         {
             base.ApplySideEffects(simulation, building);
