+++ Assets/Scripts/Controllers/BattleScene/BattleMenuDebugController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ff2ba37d6d769b541a63cea7b12b76dc
+guid: be65745807bd6dd4cbe143dacf6f04c8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
