+++ Assets/Scripts/Controllers/CitybuildingScene.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 776d3e60988841843acdf438e63be93f
-PrefabImporter:
+guid: c5662e5bbbdff8c45b3c455ba4924500
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
