+++ Assets/Scripts/DisplayObjectData/DisplayAbilityData.cs
@@ -1,4 +1,5 @@
 using Controllers;
+using Controllers.BattleScene;
 using ScriptableObjects;
 using TMPro;
 using UnityEngine;
