--- Assets/Scripts/Grid/GridBuildingSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c9db390256254e5488124a0e9d830eb5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
