--- Assets/Scripts/Grid/GridCell.cs
@@ -1,12 +1,0 @@
-namespace Grid
-{
-    public class GridCell
-    {
-        public GridCell(int gridCellType)
-        {
-            Type = gridCellType;
-        }
-
-        public int Type { get; set; }
-    }
-}
