--- Assets/Scripts/Grid/GridCell.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 365f945bd4e24341b97a2a9b48e45b04
-timeCreated: 1646642477
