--- Assets/Scripts/Grid/GridSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 86cda62f4fd51ba478346ddcddc5f67d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
