--- Assets/Scripts/Grid/GridTile.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 7c5aab5b1c2b43a6b146e0a26385c1a5
-timeCreated: 1646643822
